rename utils i testmappe -> domeneUtils.ts
