+++ src/test/domeneUtils.ts
